// src/models/enums.ts
/**
 * As cores do Núcleo de Mana, representando saltos qualitativos em poder.
 */
export enum ManaCoreColor {
    Red = "Vermelho",
    Orange = "Laranja",
    Yellow = "Amarelo",
    Cyan = "Ciano",
    Blue = "Azul",
    Violet = "Violeta",
    White = "Branco",
}

/**
 * Os níveis de refinamento corporal, um pré-requisito para avançar o núcleo.
 */
export enum BodyPurity {
    Basic = "Básica",
    Refined = "Refinada",
    Strengthened = "Fortalecida",
    Pure = "Pura",
    Transcendental = "Transcendental",
}

/**
 * Os seis blocos de construção fundamentais de todos os efeitos mágicos.
 */
export enum Element {
    Fire = "Fogo",
    Water = "Água",
    Air = "Ar",
    Earth = "Terra",
    Light = "Luz",
    Darkness = "Escuridão",
}

/**
 * As oito disciplinas que descrevem como os elementos são aplicados.
 */
export enum School {
    Abjuration = "Abjuração",
    Conjuration = "Conjuração",
    Divination = "Adivinhação",
    Enchantment = "Encantamento",
    Evocation = "Evocação",
    Illusion = "Ilusão",
    Necromancy = "Necromancia",
    Transmutation = "Transmutação",
}